Move plus voting logic
